ai-agent-demo-factory-backend/crawl4ai-agent/requirements.txt
# SmartMirrorAgent Requirements

# Core dependencies
crawl4ai>=0.3.0
aiohttp>=3.8.0
asyncio
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Quality monitoring and assessment
aiofiles>=23.0.0
pathlib2>=2.3.0

# Data handling
pandas>=1.5.0
numpy>=1.24.0

# Learning and memory
# sqlite3 is built-in to Python
json5>=0.9.0

# Web crawling enhancement
selenium>=4.0.0
playwright>=1.40.0
playwright-stealth>=1.0.6
requests>=2.28.0

# Content processing
markdown>=3.4.0
readability-lxml>=0.8.0
python-magic>=0.4.0

# Asset handling
Pillow>=9.0.0
cssutils>=2.7.0

# Logging and monitoring
loguru>=0.7.0
rich>=13.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0